# specs/001-ai-resource-hub/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI Resource Hub (Feature Contract)
  version: 0.1.0
  description: |
    Contract additions for feature 001-ai-resource-hub.

    Note: This is a feature-level contract intended to guide implementation.
    The repo's canonical OpenAPI remains the running backend at /api/v1/openapi.json.
servers:
  - url: /api/v1
...
paths:
  /resources:
    get:
      tags: [resources]
      summary: List published resources
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: Keyword search across title+description
        - in: query
          name: type
          schema: { type: string }
          description: Filter by resource type
        - in: query
          name: is_published
          schema: { type: boolean }
          description: |
            Filter by publication status.
            Note: unauthenticated callers always receive published resources.
        - in: query
          name: skip
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcesPublic'
    post:
      tags: [admin]
      summary: Create a resource (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceCreate'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePublic'
        '409':
          description: Conflict (duplicate destination_url)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /resources/{id}:
    get:
      tags: [resources]
      summary: Get resource by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePublic'
    put:
      tags: [admin]
      summary: Update resource (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePublic'
    delete:
      tags: [admin]
      summary: Delete resource (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /resources/{id}/like:
    post:
      tags: [resource-reactions]
      summary: Like resource (toggle on)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionState'
    delete:
      tags: [resource-reactions]
      summary: Remove like (toggle off)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionState'
  /resources/{id}/favorite:
    post:
      tags: [resource-reactions]
      summary: Favorite resource (toggle on)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionState'
    delete:
      tags: [resource-reactions]
      summary: Remove favorite (toggle off)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionState'
  /me/favorites:
    get:
      tags: [resource-reactions]
      summary: List my favorited resources
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: skip
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcesPublic'
  /resources/{id}/comments:
    get:
      tags: [resource-comments]
      summary: List comments for a resource
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentsPublic'
    post:
      tags: [resource-comments]
      summary: Create comment on a resource
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentPublic'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Message:
      type: object
      properties:
        message: { type: string }
      required: [message]
    ResourceType:
      type: string
      enum: [github_repo, tool, website, article, other]
    ResourceBase:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        type: { $ref: '#/components/schemas/ResourceType' }
        destination_url: { type: string, format: uri }
      required: [title, description, type, destination_url]
    ResourceCreate:
      allOf:
        - $ref: '#/components/schemas/ResourceBase'
    ResourceUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        type: { $ref: '#/components/schemas/ResourceType' }
        destination_url: { type: string, format: uri }
    ResourcePublic:
      allOf:
        - $ref: '#/components/schemas/ResourceBase'
        - type: object
          properties:
            id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
            like_count: { type: integer }
            favorite_count: { type: integer }
            comment_count: { type: integer }
          required: [id, created_at, updated_at]
    ResourcesPublic:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/ResourcePublic' }
        count: { type: integer }
      required: [data, count]
...
    CommentCreate:
      type: object
      properties:
        body: { type: string }
      required: [body]
...
    CommentPublic:
      type: object
      properties:
        id: { type: string, format: uuid }
        body: { type: string }
        author_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, body, author_id, created_at, updated_at]
    CommentsPublic:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/CommentPublic' }
        count: { type: integer }
      required: [data, count]
    ReactionState:
      type: object
      properties:
        active: { type: boolean }
      required: [active]